# docs/public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Asmblyr Core HTTP API",
    "version": "0.0.0",
    "description": "Все статические маршруты Core и их требования доступа. Часть операций имеет только route-only описание; спецификация не является полной схемой для генерации клиента. OAuth protocol и динамические маршруты плагинов описаны отдельно."
  },
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "Локальный Core"
    }
  ],
  "paths": {
...
    "/files": {
      "get": {
        "tags": [
          "files"
        ],
        "operationId": "get__files",
        "summary": "GET /files",
        "description": "Человек + files/read либо update; superuser bypass. Общая библиотека команды. Схемы тела и ответа пока не детализированы; см. реализацию обработчика. Источник: apps/core/src/files/routes.ts",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [],
        "responses": {
          "default": {
            "description": "Контракт ответа смотрите в обработчике; эта операция описана на уровне маршрута и доступа."
          }
        },
        "x-contract-level": "route-only",
        "x-source": "apps/core/src/files/routes.ts"
      },
      "post": {
        "tags": [
          "files"
        ],
        "operationId": "post__files",
        "summary": "POST /files",
        "description": "Человек + files/update; superuser bypass. Загрузка до 25 MiB, валидация метаданных, защита используемых ссылок. Схемы тела и ответа пока не детализированы; см. реализацию обработчика. Источник: apps/core/src/files/routes.ts",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [],
        "responses": {
          "default": {
            "description": "Контракт ответа смотрите в обработчике; эта операция описана на уровне маршрута и доступа."
          }
        },
        "x-contract-level": "route-only",
        "x-source": "apps/core/src/files/routes.ts"
      }
    },
    "/files/{id}": {
      "delete": {
        "tags": [
          "files"
        ],
        "operationId": "delete__files_id",
        "summary": "DELETE /files/:id",
        "description": "Человек + files/update; superuser bypass. Загрузка до 25 MiB, валидация метаданных, защита используемых ссылок. Схемы тела и ответа пока не детализированы; см. реализацию обработчика. Источник: apps/core/src/files/routes.ts",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "Контракт ответа смотрите в обработчике; эта операция описана на уровне маршрута и доступа."
          }
        },
        "x-contract-level": "route-only",
        "x-source": "apps/core/src/files/routes.ts"
      },
      "get": {
        "tags": [
          "files"
        ],
        "operationId": "get__files_id",
        "summary": "GET /files/:id",
        "description": "Активный principal: superuser, человек с files/read/update либо доступная ссылка в разрешённом поле записи; чужие файлы скрыты. Схемы тела и ответа пока не детализированы; см. реализацию обработчика. Источник: apps/core/src/files/routes.ts",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "Контракт ответа смотрите в обработчике; эта операция описана на уровне маршрута и доступа."
          }
        },
        "x-contract-level": "route-only",
        "x-source": "apps/core/src/files/routes.ts"
      },
      "patch": {
        "tags": [
          "files"
        ],
        "operationId": "patch__files_id",
        "summary": "PATCH /files/:id",
        "description": "Человек + files/update; superuser bypass. Загрузка до 25 MiB, валидация метаданных, защита используемых ссылок. Схемы тела и ответа пока не детализированы; см. реализацию обработчика. Источник: apps/core/src/files/routes.ts",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "Контракт ответа смотрите в обработчике; эта операция описана на уровне маршрута и доступа."
          }
        },
        "x-contract-level": "route-only",
        "x-source": "apps/core/src/files/routes.ts"
      }
    },
    "/files/{id}/content": {
      "get": {
        "tags": [
          "files"
        ],
        "operationId": "get__files_id_content",
        "summary": "GET /files/:id/content",
        "description": "Активный principal: superuser, человек с files/read/update либо доступная ссылка в разрешённом поле записи; чужие файлы скрыты. Схемы тела и ответа пока не детализированы; см. реализацию обработчика. Источник: apps/core/src/files/routes.ts",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "Контракт ответа смотрите в обработчике; эта операция описана на уровне маршрута и доступа."
          }
        },
        "x-contract-level": "route-only",
        "x-source": "apps/core/src/files/routes.ts"
      }
    },
    "/files/{id}/events": {
      "get": {
        "tags": [
          "files"
        ],
        "operationId": "get__files_id_events",
        "summary": "GET /files/:id/events",
        "description": "Человек + files/read либо update; superuser bypass. Общая библиотека команды. Схемы тела и ответа пока не детализированы; см. реализацию обработчика. Источник: apps/core/src/files/routes.ts",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "Контракт ответа смотрите в обработчике; эта операция описана на уровне маршрута и доступа."
          }
        },
        "x-contract-level": "route-only",
        "x-source": "apps/core/src/files/routes.ts"
      }
    },
    "/files/resolve": {
      "get": {
        "tags": [
          "files"
        ],
        "operationId": "get__files_resolve",
        "summary": "GET /files/resolve",
        "description": "Активный principal: superuser, человек с files/read/update либо доступная ссылка в разрешённом поле записи; чужие файлы скрыты. Схемы тела и ответа пока не детализированы; см. реализацию обработчика. Источник: apps/core/src/files/routes.ts",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [],
        "responses": {
          "default": {
            "description": "Контракт ответа смотрите в обработчике; эта операция описана на уровне маршрута и доступа."
          }
        },
        "x-contract-level": "route-only",
        "x-source": "apps/core/src/files/routes.ts"
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Непрозрачный access token пользователя или сервиса. Разделы настроек доступны только человеку."
      }
    }
  },
...
}